refactor protobuf method structure
