+++ webclient/src/websocket/commands/session/disconnect.ts
@@ -1,0 +1,5 @@
+import webClient from "../../WebClient";
+
+export function disconnect(): void {
+  webClient.disconnect();
+}
