+++ webclient/src/websocket/commands/session/updateStatus.ts
@@ -1,0 +1,6 @@
+import { StatusEnum } from "types"
+import webClient from "../../WebClient"
+
+export function updateStatus(status: StatusEnum, description: string): void {
+  webClient.updateStatus(status, description);
+}
