+++ webclient/src/websocket/events/room/listGames.ts
@@ -1,0 +1,7 @@
+import { RoomPersistence } from "../../persistence";
+import { ListGamesData, RoomEvent } from "./interfaces";
+
+export function listGames({ gameList }: ListGamesData, { roomEvent }: RoomEvent) {
+  const { roomId } = roomEvent;
+  RoomPersistence.updateGames(roomId, gameList);
+}
