+++ webclient/src/websocket/events/room/roomSay.ts
@@ -1,0 +1,9 @@
+import { Message } from "types";
+
+import { RoomPersistence } from "../../persistence";
+import { RoomEvent } from "./interfaces";
+
+export function roomSay(message: Message, { roomEvent }: RoomEvent) {
+  const { roomId } = roomEvent;
+  RoomPersistence.addMessage(roomId, message);
+}
