+++ webclient/src/websocket/events/session/notifyUser.ts
@@ -1,0 +1,3 @@
+export function notifyUser(payload) {
+  console.info('Event_NotifyUser', payload);
+}
