+++ webclient/src/websocket/events/session/playerPropertiesChanges.ts
@@ -1,0 +1,3 @@
+export function playerPropertiesChanges(payload) {
+  console.info('Event_PlayerPropertiesChanges', payload);
+}
