+++ webclient/src/websocket/events/session/serverMessage.ts
@@ -1,0 +1,6 @@
+import { SessionPersistence } from "../../persistence";
+import { ServerMessageData } from "./interfaces";
+
+export function serverMessage({ message }: ServerMessageData) {
+  SessionPersistence.serverMessage(message);
+}
