+++ webclient/src/websocket/events/session/serverShutdown.ts
@@ -1,0 +1,3 @@
+export function serverShutdown(payload) {
+  console.info('Event_ServerShutdown', payload);
+}
