+++ webclient/src/websocket/events/session/userJoined.ts
@@ -1,0 +1,6 @@
+import { SessionPersistence } from "../../persistence";
+import { UserJoinedData } from "./interfaces";
+
+export function userJoined({ userInfo }: UserJoinedData) {
+  SessionPersistence.userJoined(userInfo);
+}
