+++ webclient/src/websocket/events/session/userLeft.ts
@@ -1,0 +1,6 @@
+import { SessionPersistence } from "../../persistence";
+import { UserLeftData } from "./interfaces";
+
+export function userLeft({ name }: UserLeftData) {
+  SessionPersistence.userLeft(name);
+}
