+++ webclient/src/websocket/events/session/userMessage.ts
@@ -1,0 +1,3 @@
+export function userMessage(payload) {
+  console.info('Event_UserMessage', payload);
+}
